@@ -8,22 +8,27 @@
     "build": "yarn lingui:extract && next build",
     "start": "next start",
     "lingui:extract": "lingui extract --clean",
-    "test": "yarn build"
+    "test": "yarn build",
+    "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@lingui/core": "^4.8.0-next.0",
-    "@lingui/react": "^4.8.0-next.0",
-    "next": "13.5.6",
+    "@lingui/core": "^4.11.0",
+    "@lingui/react": "^4.11.0",
+    "negotiator": "^0.6.3",
+    "next": "^14.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@lingui/cli": "^4.8.0-next.0",
-    "@lingui/loader": "^4.8.0-next.0",
-    "@lingui/swc-plugin": "4.0.6",
-    "@types/react": "^18.0.14",
-    "eslint": "8.35.0",
-    "eslint-config-next": "12.3.4",
-    "typescript": "^4.7.4"
+    "@lingui/cli": "^4.11.0",
+    "@lingui/loader": "^4.11.0",
+    "@lingui/macro": "^4.11.0",
+    "@lingui/swc-plugin": "4.0.7",
+    "@types/negotiator": "^0.6.3",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^14.2.3",
+    "typescript": "^5.4.5"
   }
 }
